1. BASE_URL sequence change:
changed where BASE_URL is determined. Logic and value of base_url is now in constants.php.
This is because constants.php loads first in CI so I can't call base_url() inside constants.php if needed.
Scenarios when it is needed are when we use constant image directory to be put in IMG SRC attribute.
BASE_URL constant is then passed in config here.

2. Uploading image receipts now supported both in Create and Edit PR.

3. When editing PR, PRNum is set to be not editable